[feat]: contact form and refacto url

+++ routes/web.php
@@ -1,7 +1,8 @@
 <?php
 
+use App\Http\Controllers\admin\AdminContactController;
+use App\Http\Controllers\admin\DashboardController;
 use App\Http\Controllers\ContactController;
-use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\GuestController;
 use App\Http\Controllers\PostController;
 use Illuminate\Support\Facades\Route;
@@ -26,16 +27,21 @@ Route::middleware(['auth', 'verified'])->group(function () {
 
     // Groupe spécifique pour les Posts avec préfixe d'URL et de Nom
     Route::prefix('dashboard')->group(function () {
+
         Route::controller(PostController::class)->group(function () {
-            Route::get('/post/view', 'list')->name('post.list');
-            Route::get('/post/add', 'create')->name('post.create');
-            Route::patch('/post/{id}', 'update')->name('post.update');
-            Route::post('/store', 'store')->name('post.store');
-            Route::delete('/post/{id}', 'delete')->name('post.delete');
+            Route::get('/post/view', 'list')->name('admin.post.list');
+            Route::get('/post/add', 'create')->name('admin.post.create');
+            Route::patch('/post/{id}', 'update')->name('admin.post.update');
+            Route::post('/store', 'store')->name('admin.post.store');
+            Route::delete('/post/{id}', 'delete')->name('admin.post.delete');
 
             // Routes utilitaires (sans name spécifique dans ton code d'origine)
             Route::post('/upload-images', 'uploadImages');
             Route::post('/delete-image', 'deleteImage');
+        });
+
+        Route::controller(AdminContactController::class)->group(function () {
+            Route::get('/contact', 'index')->name('admin.contact');
         });
     });
 });
